+++ lib/domain/model/photo.dart
@@ -2,12 +2,12 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'photo.g.dart';
-part 'photo.freezed.dart';
 
+part 'photo.freezed.dart';
 
 @freezed
 class Photo with _$Photo {
-   factory Photo({
+  factory Photo({
     required int id,
     required String tags,
     required String previewURL,
